+++ experimental/packages/opentelemetry-sdk-metrics-base/src/export/MetricExporter.ts
@@ -18,7 +18,6 @@ import { AggregationTemporality } from './AggregationTemporality';
 import { ResourceMetrics } from './MetricData';
 import {
   ExportResult,
-  ExportResultCode,
 } from '@opentelemetry/core';
 import { InstrumentType } from '../InstrumentDescriptor';
 
@@ -53,31 +52,4 @@ export interface PushMetricExporter {
    * data.
    */
   shutdown(): Promise<void>;
-}
-
-export class ConsoleMetricExporter implements PushMetricExporter {
-  protected _shutdown = true;
-  private _aggregationTemporality: AggregationTemporality;
-
-  constructor(aggregationTemporality?: AggregationTemporality) {
-    this._aggregationTemporality = aggregationTemporality ?? AggregationTemporality.CUMULATIVE;
-  }
-
-  export(metrics: ResourceMetrics, resultCallback: (result: ExportResult) => void) {
-    return resultCallback({
-      code: ExportResultCode.FAILED,
-      error: new Error('Method not implemented')
-    });
-  }
-
-  selectAggregationTemporality(_instrumentType: InstrumentType) {
-    return this._aggregationTemporality;
-  }
-
-  // nothing to do
-  async forceFlush() {}
-
-  async shutdown() {
-    this._shutdown = true;
-  }
 }