+++ experimental/packages/opentelemetry-sdk-metrics-base/src/index.ts
@@ -22,6 +22,7 @@ export * from './export/MetricProducer';
 export * from './export/MetricReader';
 export * from './export/PeriodicExportingMetricReader';
 export * from './export/InMemoryMetricExporter';
+export * from './export/ConsoleMetricExporter';
 export { InstrumentDescriptor, InstrumentType } from './InstrumentDescriptor';
 export * from './Meter';
 export * from './MeterProvider';